+++ BackEnd/src/routes/user.js
@@ -6,6 +6,7 @@ const { issuanceToken } = require('../functions/signJWT');
 const express = require('express');
 const router = express.Router();
 
+const { editImage } = require('../middleware/multer.js');
 const ctrl = require('../controllers/user');
 
 // methods for user
@@ -13,7 +14,7 @@ router.post('/login', ctrl.postLogin);
 router.post('/register', ctrl.postRegister);
 
 router.get('/profile', auth, ctrl.getProfile);
-router.patch('/profile', auth, ctrl.editImage, ctrl.editProfile);
+router.patch('/profile', auth, editImage, ctrl.editProfile);
 
 // token refresh
 router.get('/token/refresh', issuanceToken);
